requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Jinja2==3.1.2
datetime==5.1
Werkzeug==2.3.7
PyMySQL==1.1.0
mysqlclient==2.1.1
firebase-admin==6.0.1
bcrypt==4.0.1
mysql-connector-python==8.0.33
pyrebase

# The 'credentials' should look something like this:
# config.py
# credentials = 'SQLALCHEMY_DATABASE_URI': 'mysql+pymysql://username:password@localhost/dbname',

